Redefine moisture as a float between 0 and 1

+++ src/model/SensorData.h
@@ -2,7 +2,7 @@
 #define SENSORDATA_H_
 
 typedef struct {
-	int moisture;
+	float moisture;
 	float temperature;
 } SensorData;
